refactor items
